refactor plugin module
